pyproject.toml
[project]
name = "newsdigest"
version = "0.1.0"
description = "A CLI tool that fetches articles from news sources and delivers a formatted digest"
requires-python = ">=3.11"
dependencies = [
    "feedparser>=6.0",
    "rich>=13.0",
    "click>=8.1",
    "pyyaml>=6.0",
    "requests>=2.31",
]

[project.scripts]
newsdigest = "newsdigest.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/newsdigest"]

[dependency-groups]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "ruff>=0.11",
    "ty>=0.0.1a1",
]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "SIM", "TCH"]
